requirements.txt
streamlit
scipy
numpy
matplotlib
sounddevice
pytest
Pillow
